Move HasIdpName class to hoauth2-demo
